Reactify - Use custom hooks and add options page
