+++ src/entries/options.tsx
@@ -1,71 +1,28 @@
 import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
+import { ThemeProvider } from '@emotion/react';
+import theme from '../common/theme';
+import CommonOptions from '../components/CommonOptions';
+import { useFolderName, useTag } from '../hooks/optionsHooks';
+import useNotification from '../hooks/useNotification';
 
 const Options = () => {
-  const [color, setColor] = useState<string>('');
-  const [status, setStatus] = useState<string>('');
-  const [like, setLike] = useState<boolean>(false);
-
-  useEffect(() => {
-    // Restores select box and checkbox state using the preferences
-    // stored in chrome.storage.
-    chrome.storage.sync.get(
-      {
-        favoriteColor: 'red',
-        likesColor: true,
-      },
-      (items) => {
-        setColor(items.favoriteColor);
-        setLike(items.likesColor);
-      }
-    );
-  }, []);
-
-  const saveOptions = () => {
-    // Saves options to chrome.storage.sync.
-    chrome.storage.sync.set(
-      {
-        favoriteColor: color,
-        likesColor: like,
-      },
-      () => {
-        // Update status to let user know options were saved.
-        setStatus('Options saved.');
-        const id = setTimeout(() => {
-          setStatus('');
-        }, 1000);
-        return () => clearTimeout(id);
-      }
-    );
-  };
+  const [folderName, updateFolderName] = useFolderName();
+  const [tag, updateTag] = useTag();
+  const [, notify, NotificationWrapper] = useNotification();
 
   return (
-    <>
-      <div>
-        Favorite color:{' '}
-        <select
-          value={color}
-          onChange={(event) => setColor(event.target.value)}
-        >
-          <option value="red">red</option>
-          <option value="green">green</option>
-          <option value="blue">blue</option>
-          <option value="yellow">yellow</option>
-        </select>
-      </div>
-      <div>
-        <label>
-          <input
-            type="checkbox"
-            checked={like}
-            onChange={(event) => setLike(event.target.checked)}
-          />
-          I like colors.
-        </label>
-      </div>
-      <div>{status}</div>
-      <button onClick={saveOptions}>Save</button>
-    </>
+    <ThemeProvider theme={theme}>
+      <CommonOptions
+        handleClose={null}
+        tag={tag}
+        updateTag={updateTag}
+        folderName={folderName}
+        updateFolderName={updateFolderName}
+        notify={notify}
+      />
+      <NotificationWrapper />
+    </ThemeProvider>
   );
 };
